+++ storage/balance.go
@@ -1,0 +1,154 @@
+package storage
+
+import (
+	"context"
+	"encoding/binary"
+	"errors"
+	"fmt"
+
+	"github.com/AnomalyFi/hypersdk/codec"
+	"github.com/AnomalyFi/hypersdk/state"
+	tconsts "github.com/AnomalyFi/nodekit-seq/consts"
+	"github.com/ava-labs/avalanchego/database"
+	smath "github.com/ava-labs/avalanchego/utils/math"
+)
+
+// [accountPrefix] + [address]
+func BalanceKey(pk codec.Address) (k []byte) {
+	k = balanceKeyPool.Get().([]byte)
+	k[0] = balancePrefix
+	copy(k[1:], pk[:])
+	binary.BigEndian.PutUint16(k[1+codec.AddressLen:], BalanceChunks)
+	return
+}
+
+// If locked is 0, then account does not exist
+func GetBalance(
+	ctx context.Context,
+	im state.Immutable,
+	addr codec.Address,
+) (uint64, error) {
+	key, bal, _, err := getBalance(ctx, im, addr)
+	balanceKeyPool.Put(key)
+	return bal, err
+}
+
+func getBalance(
+	ctx context.Context,
+	im state.Immutable,
+	addr codec.Address,
+) ([]byte, uint64, bool, error) {
+	k := BalanceKey(addr)
+	bal, exists, err := innerGetBalance(im.GetValue(ctx, k))
+	return k, bal, exists, err
+}
+
+// Used to serve RPC queries
+func GetBalanceFromState(
+	ctx context.Context,
+	f ReadState,
+	addr codec.Address,
+) (uint64, error) {
+	k := BalanceKey(addr)
+	values, errs := f(ctx, [][]byte{k})
+	bal, _, err := innerGetBalance(values[0], errs[0])
+	balanceKeyPool.Put(k)
+	return bal, err
+}
+
+func innerGetBalance(
+	v []byte,
+	err error,
+) (uint64, bool, error) {
+	if errors.Is(err, database.ErrNotFound) {
+		return 0, false, nil
+	}
+	if err != nil {
+		return 0, false, err
+	}
+	return binary.BigEndian.Uint64(v), true, nil
+}
+
+func SetBalance(
+	ctx context.Context,
+	mu state.Mutable,
+	addr codec.Address,
+	balance uint64,
+) error {
+	k := BalanceKey(addr)
+	return setBalance(ctx, mu, k, balance)
+}
+
+func setBalance(
+	ctx context.Context,
+	mu state.Mutable,
+	key []byte,
+	balance uint64,
+) error {
+	return mu.Insert(ctx, key, binary.BigEndian.AppendUint64(nil, balance))
+}
+
+func DeleteBalance(
+	ctx context.Context,
+	mu state.Mutable,
+	addr codec.Address,
+) error {
+	return mu.Remove(ctx, BalanceKey(addr))
+}
+
+func AddBalance(
+	ctx context.Context,
+	mu state.Mutable,
+	addr codec.Address,
+	amount uint64,
+	create bool,
+) error {
+	key, bal, exists, err := getBalance(ctx, mu, addr)
+	if err != nil {
+		return err
+	}
+	// Don't add balance if account doesn't exist. This
+	// can be useful when processing fee refunds.
+	if !exists && !create {
+		return nil
+	}
+	nbal, err := smath.Add64(bal, amount)
+	if err != nil {
+		return fmt.Errorf(
+			"%w: could not add balance (bal=%d, addr=%v, amount=%d)",
+			ErrInvalidBalance,
+			bal,
+			codec.MustAddressBech32(tconsts.HRP, addr),
+			amount,
+		)
+	}
+	return setBalance(ctx, mu, key, nbal)
+}
+
+func SubBalance(
+	ctx context.Context,
+	mu state.Mutable,
+	addr codec.Address,
+	amount uint64,
+) error {
+	key, bal, _, err := getBalance(ctx, mu, addr)
+	if err != nil {
+		return err
+	}
+	nbal, err := smath.Sub(bal, amount)
+	if err != nil {
+		return fmt.Errorf(
+			"%w: could not subtract balance (bal=%d, addr=%v, amount=%d)",
+			ErrInvalidBalance,
+			bal,
+			codec.MustAddressBech32(tconsts.HRP, addr),
+			amount,
+		)
+	}
+	if nbal == 0 {
+		// If there is no balance left, we should delete the record instead of
+		// setting it to 0.
+		return mu.Remove(ctx, key)
+	}
+	return setBalance(ctx, mu, key, nbal)
+}
