+++ storage/transaction.go
@@ -1,0 +1,68 @@
+package storage
+
+import (
+	"context"
+	"encoding/binary"
+	"errors"
+
+	"github.com/AnomalyFi/hypersdk/consts"
+	"github.com/AnomalyFi/hypersdk/fees"
+	"github.com/ava-labs/avalanchego/database"
+	"github.com/ava-labs/avalanchego/ids"
+)
+
+// [txPrefix] + [txID]
+func TxKey(id ids.ID) (k []byte) {
+	k = make([]byte, 1+ids.IDLen)
+	k[0] = txPrefix
+	copy(k[1:], id[:])
+	return
+}
+
+func StoreTransaction(
+	_ context.Context,
+	db database.KeyValueWriter,
+	id ids.ID,
+	t int64,
+	success bool,
+	units fees.Dimensions,
+	fee uint64,
+) error {
+	k := TxKey(id)
+	v := make([]byte, consts.Uint64Len+1+fees.DimensionsLen+consts.Uint64Len)
+	binary.BigEndian.PutUint64(v, uint64(t))
+	if success {
+		v[consts.Uint64Len] = successByte
+	} else {
+		v[consts.Uint64Len] = failureByte
+	}
+	copy(v[consts.Uint64Len+1:], units.Bytes())
+	binary.BigEndian.PutUint64(v[consts.Uint64Len+1+fees.DimensionsLen:], fee)
+	return db.Put(k, v)
+}
+
+func GetTransaction(
+	_ context.Context,
+	db database.KeyValueReader,
+	id ids.ID,
+) (bool, int64, bool, fees.Dimensions, uint64, error) {
+	k := TxKey(id)
+	v, err := db.Get(k)
+	if errors.Is(err, database.ErrNotFound) {
+		return false, 0, false, fees.Dimensions{}, 0, nil
+	}
+	if err != nil {
+		return false, 0, false, fees.Dimensions{}, 0, err
+	}
+	t := int64(binary.BigEndian.Uint64(v))
+	success := true
+	if v[consts.Uint64Len] == failureByte {
+		success = false
+	}
+	d, err := fees.UnpackDimensions(v[consts.Uint64Len+1 : consts.Uint64Len+1+fees.DimensionsLen])
+	if err != nil {
+		return false, 0, false, fees.Dimensions{}, 0, err
+	}
+	fee := binary.BigEndian.Uint64(v[consts.Uint64Len+1+fees.DimensionsLen:])
+	return true, t, success, d, fee, nil
+}
